Makes the not-zen version more zen
This is the first thing that I normally edit in chrome with the "inspect" feature: change the page's width to 90% rather than 50%, else the website is unusable. By the way, there is also two errors in your twitter link about "Report or Sugest a feature", try clicking it. "Suggest" should take two "g" and the link is broken.